Remove unnecessary float32 cast

+++ torchtitan/models/common/moe/moe.py
@@ -260,10 +260,11 @@ class TokenChoiceTopKRouter(nn.Module):
             scores = self.gate(x)
 
         # By default, sigmoid or softmax is performed in float32 to avoid loss explosion
+        # scored is already float32 from the autocast above. 
         if self.score_func == "sigmoid":
-            scores = torch.sigmoid(scores.to(torch.float32))
+            scores = torch.sigmoid(scores)
         elif self.score_func == "softmax":
-            scores = F.softmax(scores.to(torch.float32), dim=1)
+            scores = F.softmax(scores, dim=1)
         else:
             raise NotImplementedError(f"Unknown score function {self.score_func}")
